+++ resources/views/layouts/app.blade.php
@@ -14,6 +14,5 @@
     <body data-theme="{{ config('sqlite-manager.ui.theme', 'auto') }}">
         <main>{{ $slot }}</main>
         @livewireScripts
-        <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"></script>
     </body>
 </html>